+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.4-wip
 
-- Add support for parsing Vertex AI specific fields in CountTokensResponse.
+- Add support for parsing Vertex AI specific fields in `CountTokensResponse`.
+- Add named constructors on `Schema` for each value type.
 
 ## 0.3.3